requirements/runtime.txt
etcd3==0.12.0
google-auth==1.21.2
grpcio
grpcio-tools
kubernetes==11.0.0
numpy
omegaconf==2.0.0
requests==2.24.0
tqdm==4.32.1